having issues with awaiting event handler
created tasks in stock application main. Each add stock is treated as a task which should complete before the other. i converetd addstock to public async but it needs to assign an await to something. It wont let me do it to the eventhandler. idk it runs but its running like it did before.